examples/agent_observability/OpenTelemetry-NewRelic/newrelic-otlp/requirements.txt
boto3>=1.28.0
opentelemetry-exporter-otlp>=1.24.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
googleapis-common-protos==1.63.0
idna==3.7
importlib-metadata==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
opentelemetry-api==1.24.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
protobuf==4.25.3
requests==2.32.3
setuptools==70.0.0
typing_extensions==4.12.0
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.1
